refactor(UserReservationGeneralService): 과도한 래핑 제거
