+++ src/main/java/roomescape/service/facade/UserReservationGeneralService.java
@@ -1,13 +1,10 @@
 package roomescape.service.facade;
 
 import java.time.LocalDate;
-import java.util.List;
 
 import org.springframework.stereotype.Service;
 
 import roomescape.controller.dto.CreateReservationResponse;
-import roomescape.controller.dto.FindMyReservationResponse;
-import roomescape.domain.member.Member;
 import roomescape.service.PaymentService;
 import roomescape.service.UserReservationService;
 
@@ -26,17 +23,5 @@ public class UserReservationGeneralService {
                                              Long memberId, LocalDate date, Long timeId, Long themeId) {
         paymentService.pay(orderId, amount, paymentKey);
         return userReservationService.reserve(memberId, date, timeId, themeId);
-    }
-
-    public CreateReservationResponse standby(Long memberId, LocalDate date, Long timeId, Long themeId) {
-        return userReservationService.standby(memberId, date, timeId, themeId);
-    }
-
-    public void deleteStandby(Long id, Member member) {
-        userReservationService.deleteStandby(id, member);
-    }
-
-    public List<FindMyReservationResponse> findMyReservationsWithRank(Long memberId) {
-        return userReservationService.findMyReservationsWithRank(memberId);
     }
 }
