+++ src/test/java/roomescape/service/facade/UserReservationGeneralServiceTest.java
@@ -1,15 +1,12 @@
-package roomescape.service;
+package roomescape.service.facade;
 
 import static org.assertj.core.api.Assertions.assertThat;
-import static org.assertj.core.api.Assertions.assertThatCode;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
-import static org.junit.jupiter.api.Assertions.assertAll;
 import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.Mockito.doNothing;
 
 import java.time.LocalDate;
 import java.time.LocalTime;
-import java.util.List;
 
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
@@ -25,7 +22,6 @@ import org.springframework.test.context.jdbc.Sql.ExecutionPhase;
 
 import io.restassured.RestAssured;
 import roomescape.controller.dto.CreateReservationResponse;
-import roomescape.controller.dto.FindMyReservationResponse;
 import roomescape.domain.member.Member;
 import roomescape.domain.member.Role;
 import roomescape.domain.reservation.ReservationTime;
@@ -34,11 +30,12 @@ import roomescape.global.exception.RoomescapeException;
 import roomescape.repository.MemberRepository;
 import roomescape.repository.ReservationTimeRepository;
 import roomescape.repository.ThemeRepository;
-import roomescape.service.facade.UserReservationGeneralService;
+import roomescape.service.PaymentService;
+import roomescape.service.UserReservationService;
 
 @SpringBootTest(webEnvironment = WebEnvironment.RANDOM_PORT)
 @Sql(scripts = "/truncate.sql", executionPhase = ExecutionPhase.AFTER_TEST_METHOD)
-class ReservationFacadeServiceTest {
+class UserReservationGeneralServiceTest {
 
     @LocalServerPort
     int port;
@@ -66,7 +63,6 @@ class ReservationFacadeServiceTest {
     private final Long timeId = 1L;
     private final Long themeId = 1L;
     private final Long userId = 1L;
-    private final Long adminId = 2L;
     private final String orderId = "";
     private final long amount = 1000;
     private final String paymentKey = "";
@@ -152,101 +148,6 @@ class ReservationFacadeServiceTest {
                 ).isInstanceOf(RoomescapeException.class)
                         .hasMessage("과거 예약을 추가할 수 없습니다.");
             }
-        }
-    }
-
-    @Nested
-    @DisplayName("예약 대기 생성")
-    class StandBy {
-        @DisplayName("성공: 예약 대기")
-        @Test
-        void standby() {
-            CreateReservationResponse saved = reservationFacadeService.standby(userId, date, timeId, themeId);
-            assertThat(saved.id()).isEqualTo(1L);
-        }
-
-        @DisplayName("실패: 본인의 예약에 대기를 걸 수 없다.")
-        @Test
-        void standby_CantReserveAndThenStandbyForTheSameReservation() {
-            reservationFacadeService.reserve(orderId, amount, paymentKey, userId, date, timeId, themeId);
-
-            assertThatThrownBy(() -> reservationFacadeService.standby(userId, date, timeId, themeId))
-                    .isInstanceOf(RoomescapeException.class)
-                    .hasMessage("이미 예약하셨습니다. 대기 없이 이용 가능합니다.");
-        }
-
-        @DisplayName("실패: 하나의 예약에 두 개 이상 대기를 걸 수 없다.")
-        @Test
-        void standby_CantStandbyMoreThanOnce() {
-            reservationFacadeService.reserve(orderId, amount, paymentKey, adminId, date, timeId, themeId);
-            reservationFacadeService.standby(userId, date, timeId, themeId);
-
-            assertThatThrownBy(() -> userReservationService.standby(userId, date, timeId, themeId))
-                    .isInstanceOf(RoomescapeException.class)
-                    .hasMessage("이미 대기중인 예약입니다.");
-        }
-    }
-
-    @Nested
-    @DisplayName("예약/예약 대기 조회")
-    class findMyReservationsWithRank {
-        @DisplayName("성공: 특정 멤버가 예약한 예약 및 예약대기 목록 조회")
-        @Test
-        void findAllWithRankByMemberId() {
-            reservationFacadeService.reserve(orderId, amount, paymentKey, adminId, LocalDate.parse("2060-01-01"), timeId, themeId);
-            reservationFacadeService.standby(userId, LocalDate.parse("2060-01-01"), timeId, themeId);
-            reservationFacadeService.reserve(orderId, amount, paymentKey, userId, LocalDate.parse("2060-01-02"), timeId, themeId);
-            reservationFacadeService.reserve(orderId, amount, paymentKey, userId, LocalDate.parse("2060-01-03"), timeId, themeId);
-
-            List<FindMyReservationResponse> reservations = userReservationService.findMyReservationsWithRank(userId);
-
-            assertAll(
-                    () -> assertThat(reservations)
-                            .extracting(FindMyReservationResponse::id)
-                            .containsExactly(2L, 3L, 4L),
-                    () -> assertThat(reservations)
-                            .extracting(FindMyReservationResponse::rank)
-                            .containsExactly(1L, 0L, 0L)
-            );
-        }
-    }
-
-    @Nested
-    @DisplayName("예약 대기 삭제")
-    class DeleteStandBy {
-        @DisplayName("성공: 일반유저는 본인의 예약대기를 삭제할 수 있다.")
-        @Test
-        void deleteStandby() {
-            reservationFacadeService.reserve(orderId, amount, paymentKey, adminId, date, timeId, themeId);
-            reservationFacadeService.standby(userId, date, timeId, themeId);
-            Member user = memberRepository.findById(userId).get();
-
-            assertThatCode(() -> userReservationService.deleteStandby(2L, user))
-                    .doesNotThrowAnyException();
-        }
-
-        @DisplayName("실패: 일반유저는 타인의 예약대기를 삭제할 수 없다.")
-        @Test
-        void deleteStandby_ReservedByOther() {
-            reservationFacadeService.reserve(orderId, amount, paymentKey, userId, date, timeId, themeId);
-            reservationFacadeService.standby(adminId, date, timeId, themeId);
-            Member user = memberRepository.findById(userId).get();
-
-            assertThatThrownBy(() -> userReservationService.deleteStandby(2L, user))
-                    .isInstanceOf(RoomescapeException.class)
-                    .hasMessage("자신의 예약만 삭제할 수 있습니다.");
-        }
-
-        @DisplayName("실패: 예약대기 삭제 메서드로 예약을 삭제할 수 없다.")
-        @Test
-        void deleteStandby_Cannot_Delete_Reserved() {
-            reservationFacadeService.reserve(orderId, amount, paymentKey, userId, date, timeId, themeId);
-            reservationFacadeService.standby(adminId, date, timeId, themeId);
-            Member user = memberRepository.findById(userId).get();
-
-            assertThatThrownBy(() -> userReservationService.deleteStandby(1L, user))
-                    .isInstanceOf(RoomescapeException.class)
-                    .hasMessage("예약대기가 존재하지 않아 삭제할 수 없습니다.");
         }
     }
 }
